+++ crates/solidity/src/solc/standard_json/input/settings/selection/file/mod.rs
@@ -7,8 +7,6 @@ use std::collections::HashSet;
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::solc::pipeline::Pipeline as SolcPipeline;
-
 use self::flag::Flag as SelectionFlag;
 
 /// The `solc --standard-json` output file selection.
@@ -24,7 +22,7 @@ pub struct File {
 
 impl File {
     /// Creates the selection required by our compilation process.
-    pub fn new_required(pipeline: SolcPipeline) -> Self {
+    pub fn new_required() -> Self {
         Self {
             per_file: Some(HashSet::from_iter([SelectionFlag::AST])),
             per_contract: Some(HashSet::from_iter([
@@ -32,14 +30,14 @@ impl File {
                 SelectionFlag::EVMDBC,
                 SelectionFlag::MethodIdentifiers,
                 SelectionFlag::Metadata,
-                SelectionFlag::from(pipeline),
+                SelectionFlag::Yul,
             ])),
         }
     }
 
     /// Extends the user's output selection with flag required by our compilation process.
-    pub fn extend_with_required(&mut self, pipeline: SolcPipeline) -> &mut Self {
-        let required = Self::new_required(pipeline);
+    pub fn extend_with_required(&mut self) -> &mut Self {
+        let required = Self::new_required();
 
         self.per_file
             .get_or_insert_with(HashSet::default)